+++ realm-transformer/src/main/groovy/io/realm/transformer/BytecodeModifier.groovy
@@ -37,7 +37,7 @@ class BytecodeModifier {
      * @param clazz the CtClass to add accessors to.
      */
     public static void addRealmAccessors(CtClass clazz) {
-        logger.info "  Realm: Adding accessors to ${clazz.simpleName}"
+        logger.debug "  Realm: Adding accessors to ${clazz.simpleName}"
         def methods = clazz.getDeclaredMethods()*.name
         clazz.declaredFields.each { CtField field ->
             if (!Modifier.isStatic(field.getModifiers()) && !field.hasAnnotation(Ignore.class)) {
@@ -59,7 +59,7 @@ class BytecodeModifier {
      */
     public static void useRealmAccessors(CtClass clazz, List<CtField> managedFields) {
         clazz.getDeclaredBehaviors().each { behavior ->
-            logger.info "    Behavior: ${behavior.name}"
+            logger.debug "    Behavior: ${behavior.name}"
             if (
                 (
                     behavior instanceof CtMethod &&
@@ -104,13 +104,13 @@ class BytecodeModifier {
 
         @Override
         void edit(FieldAccess fieldAccess) throws CannotCompileException {
-            logger.info "      Field being accessed: ${fieldAccess.className}.${fieldAccess.fieldName}"
+            logger.debug "      Field being accessed: ${fieldAccess.className}.${fieldAccess.fieldName}"
             def isRealmFieldAccess = managedFields.find {
                 fieldAccess.className.equals(it.declaringClass.name) && fieldAccess.fieldName.equals(it.name)
             }
             if (isRealmFieldAccess != null) {
-                logger.info "        Realm: Manipulating ${ctClass.simpleName}.${behavior.name}(): ${fieldAccess.fieldName}"
-                logger.info "        Methods: ${ctClass.declaredMethods}"
+                logger.debug "        Realm: Manipulating ${ctClass.simpleName}.${behavior.name}(): ${fieldAccess.fieldName}"
+                logger.debug "        Methods: ${ctClass.declaredMethods}"
                 def fieldName = fieldAccess.fieldName
                 if (fieldAccess.isReader()) {
                     fieldAccess.replace('$_ = $0.realmGet$' + fieldName + '();')
@@ -127,7 +127,7 @@ class BytecodeModifier {
      * @param clazz The CtClass to modify.
      */
     public static void overrideTransformedMarker(CtClass clazz) {
-        logger.info "  Realm: Marking as transformed ${clazz.simpleName}"
+        logger.debug "  Realm: Marking as transformed ${clazz.simpleName}"
         try {
             clazz.getDeclaredMethod("transformerApplied", new CtClass[0])
         } catch (NotFoundException ignored) {